+++ draw/vector/src/font/font.rs
@@ -1,5 +1,7 @@
 use crate::font::Glyph;
 use crate::geometry::Rectangle;
+use resvg::usvg::Tree;
+use std::rc::Rc;
 
 /// A font.
 #[derive(Clone, Debug)]
@@ -10,5 +12,6 @@ pub struct TTFFont {
     pub line_gap: f64,
     pub bounds: Rectangle,
     pub cached_decoded_glyphs: Vec<Option<Box<Glyph>>>,
+    pub cached_svg_images: Vec<Option<Option<Rc<Tree>>>>,
 }