+++ libs/ttf-parser/c-api/ttfparser.h
@@ -11,9 +11,9 @@
 #include <stdint.h>
 
 #define TTFP_MAJOR_VERSION 0
-#define TTFP_MINOR_VERSION 19
+#define TTFP_MINOR_VERSION 20
 #define TTFP_PATCH_VERSION 0
-#define TTFP_VERSION "0.19.0"
+#define TTFP_VERSION "0.20.0"
 
 /**
  * @brief A glyph image format.