+++ libs/ttf-parser/src/tables/mod.rs
@@ -2,6 +2,8 @@ pub mod cbdt;
 pub mod cblc;
 mod cff;
 pub mod cmap;
+pub mod colr;
+pub mod cpal;
 pub mod glyf;
 pub mod head;
 pub mod hhea;